backend/requirements.txt
# Backend dependencies
fastapi
uvicorn
pdfplumber
pytesseract
Pillow
scikit-learn
google-generativeai
python-dotenv
